refactor(extensions): address local-review findings — package splits, dedupe correctness, DRY consolidation
Addresses the seven Important + Critical findings from /devloop:local-review
on this branch. Tests still pass (1345 passed, 9 skipped).

## Critical: file/function size limits (per .ai/rules/style.md)

Both convert_agent.py (1209 lines) and sdk_override.py (946 lines)
exceeded the project's 500-line file limit, with several functions
over the 50-line limit. Split into packages along natural seams.

### convert_agent.py → convert_agent/

- ``models.py`` (62) — FileModification, ConversionStrategy,
  ValidationSummary, ConversionPlan
- ``prompts.py`` (317) — strategy + modification prompt builders +
  per-section render helpers
- ``providers.py`` (343) — call_llm + per-provider runners +
  _safe_api_call + subprocess wrapper + _AGENT_GUIDANCE loader
- ``plan.py`` (441) — parse + apply + _execute_copy/_execute_text +
  preserve-existing + supporting-files + dedupe
- ``agent.py`` (267) — run_agent orchestration + staging validation
- ``__init__.py`` (126) — back-compat re-exports

Public surface (``from kamiwaza_extensions.convert_agent import X``)
unchanged. The two test sites that monkeypatched
``convert_agent._call_anthropic`` / ``_call_openai_compatible`` now
patch via ``convert_agent.providers`` (the four call_llm patches now
target ``convert_agent.agent.call_llm`` where it's used).

### sdk_override.py → sdk_override/

- ``spec.py`` (245) — SdkOverrideSpec, resolve, validate, dist
  staleness, TS build, diagnostics
- ``classification.py`` (192) — service detectors + Dockerfile stage
  parsing
- ``compose.py`` (107) — generate_compose_override (the ENG-4413
  shell-free path)
- ``build.py`` (429) — Dockerfile overlays + dev local build patches
- ``__init__.py`` (89) — back-compat re-exports

### Function-size fixes (in the new convert_agent package)

- ``call_llm`` (107 → ~25 lines): table-drove provider dispatch via
  ``_PROVIDERS = [(name, runner), ...]``. Removed the four near-
  identical "if provider in (auto, X)" blocks.
- ``apply_plan`` (84 → ~35 lines): extracted
  ``_resolve_modification_target``, ``_execute_copy``, ``_execute_text``.
- ``build_modification_prompt`` (105 → ~70 lines): extracted
  ``_render_validation_feedback`` and ``_render_previous_plan``.
- ``run_agent`` (95 → ~75 lines): extracted ``_llm_unavailable_failure``.

## Important: real correctness bug in dedupe

``_dedupe_manual_items_against_modifications`` used naive substring
matching that produced silent false positives — a modification path of
``backend/Dockerfile`` would silently drop the legitimate manual_item
"Add Dockerfile to .gitignore" because "dockerfile" is a substring of
both. Short basenames (``go``, ``web``, ``app``) collided with normal
words like "goal", "webhook", "appears".

Fixed with whole-word regex matching plus a 4-char minimum-token
threshold. Three new regression tests cover the previously-failing
cases.

## Important: DRY (Rule of Three)

- New ``kamiwaza_extensions/monorepo.py`` is the single source of
  truth for ``MONOREPO_PARENT_DIRS``, ``MONOREPO_BARE_DIRS``, and
  ``SKIP_DIRS``. Replaces the parallel constants in
  ``app_analyzer.py`` (was ``_MONOREPO_SUBDIR_PATTERNS`` +
  ``_SKIP_DIRS``), ``extension_detector.py`` (was
  ``_MONOREPO_PARENT_DIRS``), and ``convert_agent.py`` (was
  ``_STAGING_SKIP_DIRS``, materialized from SKIP_DIRS at staging time).
- The four CLI-provider blocks were duplicated argv builders; now
  table-driven via ``_provider_*`` runners (see Critical fix above).
- The two API-provider helpers (``_call_anthropic`` and
  ``_call_openai_compatible``) shared the same try/except shape;
  extracted ``_safe_api_call(label, fn, sdk_exceptions)`` to absorb
  the wrapper, and replaced the ``except Exception`` swallowers with
  a tighter ``except (TypeError, ValueError, AttributeError, KeyError)``
  that includes the exception type in the warning.

## Documentation

``agent_guidance.md`` adds a "Vendoring Binary Dependencies" rule
pushing the LLM toward PYTHONPATH-based vendoring (and node_modules
bind/copy for Node) instead of in-image ``pip wheel`` / ``npm pack``
on vendored sources. Discovered during the smoke run: when the LLM
vendors a flat-layout Python package source whose pyproject.toml
omits ``[tool.hatch.build.targets.wheel]`` config, hatchling can't
auto-detect a wheel layout and the build fails. PYTHONPATH-based
imports sidestep the entire build step.
